+++ Net_Core/DataAccess/Interfaces/ICommunityCalendar.cs
@@ -2,6 +2,7 @@ using GrowthWare.DataAccess.Interfaces.Base;
 using GrowthWare.Framework.Models;
 using System;
 using System.Data;
+using System.Threading.Tasks;
 
 namespace GrowthWare.DataAccess.Interfaces
 {
@@ -10,12 +11,12 @@ namespace GrowthWare.DataAccess.Interfaces
         int CalendarSeqId {get; set;}
         int SecurityEntitySeqId {get; set;}
 		
-        bool DeleteEvent(int calendarEventSeqId);
+        Task<bool> DeleteEvent(int calendarEventSeqId);
 		
-        DataTable GetEvents(int functionSeqId, DateTime startDate, DateTime endDate);
+        Task<DataTable> GetEvents(int functionSeqId, DateTime startDate, DateTime endDate);
         
         DataRow GetEvent(int calendarEventSeqId);
         
-		DataRow SaveCalendarEvent(int functionSeqId, MCalendarEvent calendarEvent);
+		Task<DataRow> SaveCalendarEvent(int functionSeqId, MCalendarEvent calendarEvent);
     }
 }