+++ Net_Core/DataAccess/SQLServer/DCommunityCalendar.cs
@@ -4,6 +4,7 @@ using GrowthWare.Framework.Models;
 using System;
 using System.Data;
 using System.Data.SqlClient;
+using System.Threading.Tasks;
 
 namespace GrowthWare.DataAccess.SQLServer;
 
@@ -28,17 +29,16 @@ public class DCommunityCalendar : AbstractDBInteraction, ICommunityCalendar
     }
 #endregion
 
-    bool ICommunityCalendar.DeleteEvent(int calendarEventSeqId)
+    async Task<bool> ICommunityCalendar.DeleteEvent(int calendarEventSeqId)
     {
         this.checkValid();
         string mStoredProcedure = "[ZGWOptional].[Delete_Calendar_Event]";
-        SqlParameter[] mParameters =
-        {
+        SqlParameter[] mParameters = [
             new ("@P_CalendarEventSeqId", calendarEventSeqId),
-        };
+        ];
         try
         {
-            base.ExecuteNonQuery(mStoredProcedure, mParameters);
+            await base.ExecuteNonQueryAsync(mStoredProcedure, mParameters);
         }
         catch (System.Exception)
         {
@@ -65,19 +65,18 @@ public class DCommunityCalendar : AbstractDBInteraction, ICommunityCalendar
         }
     }
 
-    DataTable ICommunityCalendar.GetEvents(int functionSeqId, DateTime startDate, DateTime endDate)
+    async Task<DataTable> ICommunityCalendar.GetEvents(int functionSeqId, DateTime startDate, DateTime endDate)
     {
         this.checkValid();
         string mStoredProcedure = "[ZGWOptional].[Get_Calendar_Events]";
-        SqlParameter[] mParameters =
-        {
+        SqlParameter[] mParameters = [
             new ("@P_FunctionSeqId", functionSeqId),
             new ("@P_Start_Date", startDate),
             new ("@P_End_Date", endDate),
-        };
+        ];
         try
         {
-            return base.GetDataTable(mStoredProcedure, mParameters);
+            return await base.GetDataTableAsync(mStoredProcedure, mParameters);
         }
         catch (System.Exception)
         {
@@ -85,12 +84,11 @@ public class DCommunityCalendar : AbstractDBInteraction, ICommunityCalendar
         }
     }
 
-    DataRow ICommunityCalendar.SaveCalendarEvent(int functionSeqId, MCalendarEvent calendarEvent)
+    async Task<DataRow> ICommunityCalendar.SaveCalendarEvent(int functionSeqId, MCalendarEvent calendarEvent)
     {
         this.checkValid();
         string mStoredProcedure = "[ZGWOptional].[Set_Calendar_Event]";
-        SqlParameter[] mParameters =
-            {
+        SqlParameter[] mParameters = [
             new ("@P_CalendarEventSeqId", calendarEvent.Id),
             new ("@P_FunctionSeqId", functionSeqId),
             new ("@P_Title", calendarEvent.Title),
@@ -102,10 +100,10 @@ public class DCommunityCalendar : AbstractDBInteraction, ICommunityCalendar
             new ("@P_Link", calendarEvent.Link),
             new ("@P_Location", calendarEvent.Location),
             new ("@P_Added_Updated_By", calendarEvent.AddedBy),
-            };
+        ];
         try
         {
-            return base.GetDataRow(mStoredProcedure, mParameters);
+            return await base.GetDataRowAsync(mStoredProcedure, mParameters);
         }
         catch (System.Exception)
         {