+++ Net_Core/Web.Support/BaseControllers/AbstractCalendarController.cs
@@ -7,6 +7,7 @@ using GrowthWare.Framework.Models;
 using GrowthWare.Web.Support.Jwt;
 using GrowthWare.Web.Support.Utilities;
 using GrowthWare.Framework.Models.UI;
+using System.Threading.Tasks;
 
 namespace GrowthWare.Web.Support.BaseControllers;
 
@@ -24,9 +25,10 @@ public abstract class AbstractCalendarController : ControllerBase
     /// <returns></returns>
     [AllowAnonymous]
     [HttpGet("DeleteEvent")]
-    public ActionResult<bool> DeleteEvent(int calendarEventSeqId, string action)
+    public async Task<ActionResult<bool>> DeleteEvent(int calendarEventSeqId, string action)
     {
-        if (getEventSecurity(calendarEventSeqId))
+        MCalendarEvent mCalendarEvent = CalendarUtility.GetEvent(SecurityEntityUtility.CurrentProfile, calendarEventSeqId);
+        if (getEventSecurity(mCalendarEvent))
         {
             MFunctionProfile mFunctionProfile = FunctionUtility.GetProfile(action);
             if (mFunctionProfile != null)
@@ -35,7 +37,7 @@ public abstract class AbstractCalendarController : ControllerBase
                 MSecurityInfo mSecurityInfo = new(mFunctionProfile, mAccountProfile);
                 if (mSecurityInfo.MayView)
                 {
-                    return Ok(CalendarUtility.DeleteEvent(SecurityEntityUtility.CurrentProfile, calendarEventSeqId));
+                    return Ok(await CalendarUtility.DeleteEvent(SecurityEntityUtility.CurrentProfile, calendarEventSeqId));
                 }
             }
         }
@@ -82,7 +84,7 @@ public abstract class AbstractCalendarController : ControllerBase
     /// <returns></returns>
     [AllowAnonymous]
     [HttpGet("GetEvents")]
-    public ActionResult<List<MCalendarEvent>> GetEvents(string action, string startDate, string endDate)
+    public async Task<ActionResult<List<MCalendarEvent>>> GetEvents(string action, string startDate, string endDate)
     {
         /** 
           * The startDate and endDate parameters are needed because the data for a calendar is
@@ -98,7 +100,7 @@ public abstract class AbstractCalendarController : ControllerBase
             {
                 DateTime mStartDate = DateTime.Parse(startDate);
                 DateTime mEndtDate = DateTime.Parse(endDate);
-                List<MCalendarEvent> mRetVal = CalendarUtility.GetEvents(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mStartDate, mEndtDate);
+                List<MCalendarEvent> mRetVal = await CalendarUtility.GetEvents(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mStartDate, mEndtDate);
                 return Ok(mRetVal);
             }
             return StatusCode(StatusCodes.Status401Unauthorized, "The requesting account does not have the correct permissions");
@@ -118,7 +120,7 @@ public abstract class AbstractCalendarController : ControllerBase
     /// </remarks>
     [AllowAnonymous]
     [HttpPost("SaveEvent")]
-    public ActionResult<MCalendarEvent> SaveEvent(UISaveEventParameters parameters)
+    public async Task<ActionResult<MCalendarEvent>> SaveEvent(UISaveEventParameters parameters)
     {
         MAccountProfile mAccountProfile = AccountUtility.CurrentProfile;
         UISaveEventParameters mParameters = parameters; // Bad practice to alter a parameter in a method.
@@ -132,10 +134,11 @@ public abstract class AbstractCalendarController : ControllerBase
             mParameters.calendarEvent.UpdatedBy = mAccountProfile.Id;
             mParameters.calendarEvent.UpdatedDate = DateTime.Now;
         }
-        if (getEventSecurity(mParameters.calendarEvent.Id, mParameters.action) && mParameters.calendarEvent.AddedBy == mAccountProfile.Id)
+        MCalendarEvent mCalendarEvent = CalendarUtility.GetEvent(SecurityEntityUtility.CurrentProfile, mParameters.calendarEvent.Id);
+        if (getEventSecurity(mCalendarEvent, mParameters.action) && mParameters.calendarEvent.AddedBy == mAccountProfile.Id)
         {
             MFunctionProfile mFunctionProfile = FunctionUtility.GetProfile(mParameters.action);
-            MCalendarEvent mRetVal = CalendarUtility.SaveCalendarEvent(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mParameters.calendarEvent);
+            MCalendarEvent mRetVal = await CalendarUtility.SaveCalendarEvent(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mParameters.calendarEvent);
             return Ok(mRetVal);
         }
         return StatusCode(StatusCodes.Status401Unauthorized, "The requesting account does not have the correct permissions");
@@ -145,24 +148,21 @@ public abstract class AbstractCalendarController : ControllerBase
     [HttpGet("GetEventSecurity")]
     public ActionResult<Boolean> GetEventSecurity(int calendarEventSeqId)
     {
-        return Ok(getEventSecurity(calendarEventSeqId));
+        MCalendarEvent mCalendarEvent = CalendarUtility.GetEvent(SecurityEntityUtility.CurrentProfile, calendarEventSeqId);
+        return Ok(getEventSecurity(mCalendarEvent));
     }
 
-    private bool getEventSecurity(int calendarEventSeqId, string action = null)
+    private bool getEventSecurity(MCalendarEvent calendarEvent, string action = null)
     {
         bool mRetVal = false;
         MSecurityEntity mSecurityEntity = SecurityEntityUtility.CurrentProfile;
         if (mSecurityEntity != null)
         {
-            MCalendarEvent mCalendarEvent = CalendarUtility.GetEvent(mSecurityEntity, calendarEventSeqId);
             if (action == null)
             {
-                if (mCalendarEvent != null)
+                if (AccountUtility.CurrentProfile.Id == calendarEvent.AddedBy)
                 {
-                    if (AccountUtility.CurrentProfile.Id == mCalendarEvent.AddedBy)
-                    {
-                        mRetVal = true;
-                    }
+                    mRetVal = true;
                 }
             }
             else