+++ Release_Notes.md
@@ -4,6 +4,9 @@
 ### New Features
 
 ### Improvements
+- Converted database calls to async/await
+    - DCommunityCalendar Excluding:
+        - DataRow GetEvent(int calendarEventSeqId)
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure